+++ CHANGELOG.md
@@ -6,7 +6,9 @@
 * Swift Package Manager
   * Update Package.swift to include `PPRiskMagnes` as explicit target for library products that require `PayPalDataCollector`
   * _Note:_ No longer requires manual inclusion of `PayPalDataCollector` in order to use `BraintreeThreeDSecure`, `BraintreePayPal`, and `BraintreePaymentFlow`
-  
+* Local Payment Methods
+  * Add `displayName` to `BTLocalPaymentRequest`
+
 ## 4.38.0 (2021-08-24)
 * Add `offerPayLater` to `BTPayPalRequest`